refactor: move viewers to viewers dir
