+++ src/pytorch_faster_rcnn_tutorial/viewers/anchor_viewer.py
@@ -8,7 +8,7 @@ from torchvision.models.detection.transform import GeneralizedRCNNTransform
 from torchvision.ops import box_convert
 
 from pytorch_faster_rcnn_tutorial.anchor_generator import get_anchor_boxes
-from pytorch_faster_rcnn_tutorial.dataset_viewer import DatasetViewer, make_bbox_napari
+from pytorch_faster_rcnn_tutorial.viewers.dataset_viewer import DatasetViewer, make_bbox_napari
 from pytorch_faster_rcnn_tutorial.transformations import re_normalize
